bot/requirements.txt
aiogram
asyncpg
aiohttp
nats-py
pytest
pytest-asyncio
